pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-hum-crawler"
version = "0.2.0"
description = "Dynamic disaster intelligence monitoring assistant"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Fawad" }]

dependencies = [
  "APScheduler>=3.10.4",
  "beautifulsoup4>=4.12.3",
  "feedparser>=6.0.11",
  "httpx>=0.27.2",
  "pdfplumber>=0.11.4",
  "pydantic>=2.9.2",
  "pypdf>=5.1.0",
  "python-dotenv>=1.0.1",
  "sqlmodel>=0.0.22",
  "trafilatura>=1.12.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3"
]

[project.scripts]
agent-hum-crawler = "agent_hum_crawler.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
